+++ WpfApp1/PickAdditionalServices.xaml
@@ -1,0 +1,31 @@
+﻿<Window x:Class="WpfApp1.PickAdditionalServices"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:WpfApp1"
+        mc:Ignorable="d"
+        Title="PickAdditionalServices" Height="387.668" Width="869.891" Loaded="Window_Loaded">
+    <Grid>
+        <DataGrid SelectionMode="Single" IsReadOnly="True" AutoGenerateColumns="False" x:Name="additionalServicesDG" HorizontalAlignment="Left" Height="130" Margin="29,126,0,0" VerticalAlignment="Top" Width="403">
+            <DataGrid.Columns>
+                <DataGridTextColumn Header="Услуга" Binding="{Binding additional_service_name}" Width="280"/>
+                <DataGridTextColumn Header="Стоимость" Binding="{Binding monthly_fee}" Width="70"/>
+            </DataGrid.Columns>
+        </DataGrid>
+        <DataGrid x:Name="orderAddServiceDG" SelectionMode="Single" HorizontalAlignment="Left" Height="130" Margin="491,126,0,0" VerticalAlignment="Top" Width="328">
+            <DataGrid.Columns>
+                <DataGridTextColumn Header="Услуга" Binding="{Binding additional_service_name}"/>
+            </DataGrid.Columns>
+        </DataGrid>
+        <Label Content="Поиск доп. услуги" HorizontalAlignment="Left" VerticalAlignment="Top" FontSize="16" Height="31" Width="256" Margin="29,63,0,0"/>
+        <TextBox x:Name="searchAddServiceTextBox" HorizontalAlignment="Left" Height="23" Margin="28,95,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="256" KeyUp="searchTextBox_KeyUp"/>
+        <Label Content="Доп. услуги наряда" HorizontalAlignment="Left" VerticalAlignment="Top" FontSize="16" Height="31" Width="256" Margin="491,86,0,0"/>
+        <Button Content="&gt;" HorizontalAlignment="Left" Margin="441,165,0,0" VerticalAlignment="Top" Width="42" FontSize="16" Background="{x:Null}" Click="Button_Click_2"/>
+        <Button Content="&lt;" HorizontalAlignment="Left" Margin="441,201,0,0" VerticalAlignment="Top" Width="42" FontSize="16" Background="{x:Null}" Click="Button_Click_5"/>
+        <Label Content="Оплата в месяц:" HorizontalAlignment="Left" Margin="633,249,0,0" VerticalAlignment="Top" FontSize="14" RenderTransformOrigin="0.337,0.402"/>
+        <Label x:Name="addServicesTotalCostLabel" Content="" HorizontalAlignment="Left" Margin="746,249,0,0" VerticalAlignment="Top" FontSize="14" RenderTransformOrigin="0.337,0.402"/>
+        <Button Content="К заявке" HorizontalAlignment="Left" Margin="517,295,0,0" VerticalAlignment="Top" Width="118" Height="30"/>
+
+    </Grid>
+</Window>
